fixed a ccreating table errorr

+++ Backend/userDB/src/main/java/database/UserDB.java
@@ -4,8 +4,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-//public interface UserDB extends JpaRepository<User, Long>{
-public interface UserDB extends JpaRepository<User, Integer>{
+public interface UserDB extends JpaRepository<User, Long>{
+//public interface UserDB extends JpaRepository<User, Integer>{
 
 
 }
